src/mod-1-document-extract/requirements.txt
airflow-ai-sdk[openai]==0.2.0

langfuse==3.3.0

apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.16.0

python-dotenv==1.0.0
pyyaml==6.0.1
boto3==1.34.0
psycopg2-binary==2.9.9

opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0